signature_extractor_v5/requirements.txt
# requirements.txt
pandas
requests
PyMuPDF
pillow
python-dateutil

# Docling and transformers
docling
transformers
torch

# Optional for better performance
accelerate
sentencepiece